refactor: move FA events to its own module
